Revert "Manuel"
