+++ package.json
@@ -11,8 +11,6 @@
   "dependencies": {
     "@fortawesome/fontawesome-free": "^6.7.2",
     "axios": "^1.8.4",
-    "firebase": "^11.5.0",
-    "firebase-admin": "^13.2.0",
     "vue": "^3.5.13",
     "vue-router": "^4.5.0"
   },
