--- src/bbdd.json
@@ -1,172 +1,0 @@
-[
-    {
-      "id": 1,
-      "modelo": "Classic Fit",
-      "color": "Blanco",
-      "material": "Algodón 100%",
-      "marca": "Nike",
-      "genero": "Unisex",
-      "diseno": "Liso",
-      "temporada": "Todo el año",
-      "precio": 24.99,
-      "stock_por_talla": {
-        "S": 15,
-        "M": 25,
-        "L": 20,
-        "XL": 10
-      }
-    },
-    {
-      "id": 2,
-      "modelo": "Sport Pro",
-      "color": "Negro",
-      "material": "Poliéster 90%",
-      "marca": "Adidas",
-      "genero": "Hombre",
-      "diseno": "Rayas laterales",
-      "temporada": "Verano",
-      "precio": 29.99,
-      "stock_por_talla": {
-        "S": 12,
-        "M": 18,
-        "L": 22,
-        "XL": 8
-      }
-    },
-    {
-      "id": 3,
-      "modelo": "Vintage Logo",
-      "color": "Rojo",
-      "material": "Algodón 80%",
-      "marca": "Levi's",
-      "genero": "Unisex",
-      "diseno": "Estampado retro",
-      "temporada": "Otoño",
-      "precio": 22.50,
-      "stock_por_talla": {
-        "S": 10,
-        "M": 15,
-        "L": 10,
-        "XL": 5
-      }
-    },
-    {
-      "id": 4,
-      "modelo": "Ocean Breeze",
-      "color": "Azul marino",
-      "material": "Algodón orgánico",
-      "marca": "H&M",
-      "genero": "Mujer",
-      "diseno": "Degradé",
-      "temporada": "Primavera",
-      "precio": 19.99,
-      "stock_por_talla": {
-        "S": 20,
-        "M": 30,
-        "L": 15,
-        "XL": 5
-      }
-    },
-    {
-      "id": 5,
-      "modelo": "Urban Edge",
-      "color": "Gris",
-      "material": "Mezcla algodón/poliéster",
-      "marca": "Zara",
-      "genero": "Hombre",
-      "diseno": "Logo minimalista",
-      "temporada": "Todo el año",
-      "precio": 27.99,
-      "stock_por_talla": {
-        "S": 8,
-        "M": 20,
-        "L": 15,
-        "XL": 7
-      }
-    },
-    {
-      "id": 6,
-      "modelo": "Floral Dream",
-      "color": "Rosa",
-      "material": "Algodón 95%",
-      "marca": "Mango",
-      "genero": "Mujer",
-      "diseno": "Flores estampadas",
-      "temporada": "Primavera",
-      "precio": 21.99,
-      "stock_por_talla": {
-        "S": 18,
-        "M": 25,
-        "L": 12,
-        "XL": 5
-      }
-    },
-    {
-      "id": 7,
-      "modelo": "Tech Dry",
-      "color": "Negro",
-      "material": "Poliéster 100%",
-      "marca": "Under Armour",
-      "genero": "Unisex",
-      "diseno": "Costuras reflectantes",
-      "temporada": "Verano",
-      "precio": 34.99,
-      "stock_por_talla": {
-        "S": 10,
-        "M": 15,
-        "L": 10,
-        "XL": 3
-      }
-    },
-    {
-      "id": 8,
-      "modelo": "Band Tee",
-      "color": "Blanco",
-      "material": "Algodón 100%",
-      "marca": "Pull & Bear",
-      "genero": "Unisex",
-      "diseno": "Estampado de banda",
-      "temporada": "Todo el año",
-      "precio": 26.50,
-      "stock_por_talla": {
-        "S": 5,
-        "M": 12,
-        "L": 8,
-        "XL": 2
-      }
-    },
-    {
-      "id": 9,
-      "modelo": "Crop Top",
-      "color": "Amarillo",
-      "material": "Algodón 85%",
-      "marca": "Bershka",
-      "genero": "Mujer",
-      "diseno": "Liso",
-      "temporada": "Verano",
-      "precio": 18.99,
-      "stock_por_talla": {
-        "S": 22,
-        "M": 30,
-        "L": 10,
-        "XL": 4
-      }
-    },
-    {
-      "id": 10,
-      "modelo": "Boxy Fit",
-      "color": "Verde militar",
-      "material": "Algodón 100%",
-      "marca": "Uniqlo",
-      "genero": "Unisex",
-      "diseno": "Bolsillo delantero",
-      "temporada": "Otoño",
-      "precio": 23.99,
-      "stock_por_talla": {
-        "S": 10,
-        "M": 20,
-        "L": 15,
-        "XL": 8
-      }
-    }
-  ]
